requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.4
django-filter==22.1
django-htmx==1.13.0
django-phonenumber-field==7.0.1
django-widget-tweaks==1.4.12
feedparser==6.0.10
phonenumbers==8.13.3
Pillow==9.3.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
